Core - Rename CefSharp.Internals.CefPostDataWrapper to CefSharp.PostDataElement
- Make public
- Added public constructor

Related issue #1721
