+++ CefSharp.Core/CefSharp.Core.vcxproj.filters
@@ -202,9 +202,6 @@
     <ClInclude Include="Internals\CefPostDataWrapper.h">
       <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="Internals\CefPostDataElementWrapper.h">
-      <Filter>Header Files</Filter>
-    </ClInclude>
     <ClInclude Include="Internals\JavascriptCallbackProxy.h">
       <Filter>Header Files</Filter>
     </ClInclude>
@@ -323,6 +320,9 @@
       <Filter>Header Files</Filter>
     </ClInclude>
     <ClInclude Include="Internals\CefUrlRequestWrapper.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="PostDataElement.h">
       <Filter>Header Files</Filter>
     </ClInclude>
   </ItemGroup>
